refactor(app): move action components into Controls.tsx
