+++ srcs/app/index.css
@@ -2,6 +2,12 @@
 @tailwind components;
 @tailwind utilities;
 
+@layer components {
+  .control-btn {
+    @apply z-50 flex h-8 w-8 place-content-center place-items-center rounded border-[1px] border-[#505050] bg-[#2D2D2D] text-gray-400 transition-colors hover:border-[#777777] hover:bg-[#333333];
+  }
+}
+
 body {
   background-color: #999999;
   text-rendering: optimizeLegibility;
